feat: Remove unnecessary packages
