+++ package.json
@@ -25,19 +25,13 @@
     "start": "yarn watch",
     "build-start": "yarn build && yarn watch",
     "watch": "run-p watch:*",
-    "watch:utils": "yarn workspace @aviato/utils watch",
-    "watch:theme": "yarn workspace @aviato/theme watch",
     "watch:ui": "yarn workspace @aviato/ui watch",
     "watch:docs": "yarn workspace @aviato/docs dev",
     "lint": "run-p lint:*",
-    "lint:utils": "yarn workspace @aviato/utils lint",
-    "lint:theme": "yarn workspace @aviato/theme lint",
     "lint:ui": "yarn workspace @aviato/ui lint",
     "lint:docs": "yarn workspace @aviato/docs lint",
     "prebuild": "yarn lint",
     "build": "run-s build:*",
-    "build:utils": "yarn workspace @aviato/utils build",
-    "build:theme": "yarn workspace @aviato/theme build",
     "build:ui": "yarn workspace @aviato/ui build",
     "build:docs": "yarn workspace @aviato/docs build"
   },
