+++ packages/docs/pages/_document.tsx
@@ -1,5 +1,5 @@
 import NextDocument, { Html, Head, Main, NextScript } from 'next/document'
-import { getCssText } from '@aviato/ui/theme'
+import { getCssText } from '@aviato/ui'
 
 export default class Document extends NextDocument {
   render() {
