+++ packages/docs/pages/components/sections/hooks.tsx
@@ -1,7 +1,7 @@
 import DisplayComponent from '../displayComponent'
 
-import { styled } from '@aviato/ui/theme'
-import { log } from '@aviato/ui/utils'
+import { styled } from '@aviato/ui'
+import { log } from '@aviato/ui'
 import {
   Title,
   useLongPress,
