+++ packages/docs/pages/components/sections/menu.tsx
@@ -1,5 +1,5 @@
-import { styled } from '@aviato/ui/theme'
-import { log } from '@aviato/ui/utils'
+import { styled } from '@aviato/ui'
+import { log } from '@aviato/ui'
 import { Title, Menu, MenuItem } from '@aviato/ui'
 
 import DisplayComponent from '../displayComponent'
