--- packages/theme/.gitignore
@@ -1,2 +1,0 @@
-temp
-dist
