+++ packages/ui/src/index.ts
@@ -1,2 +1,4 @@
 export * from './components'
 export * from './hooks'
+export * from './theme'
+export * from './utils'
