--- packages/ui/theme/index.ts
@@ -1,7 +1,0 @@
-/**
- * Re-export theme.
- *
- * @Note
- * This negates the need to publish.
- */
-export * from '@aviato/theme'
