--- packages/ui/tsconfig.theme.json
@@ -1,14 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist/theme",
-    "rootDir": "./theme"
-  },
-  "include": [
-    "theme/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist",
-  ],
-}
