--- packages/ui/tsconfig.utils.json
@@ -1,14 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist/utils",
-    "rootDir": "./utils"
-  },
-  "include": [
-    "utils/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist",
-  ],
-}
