--- packages/ui/utils/index.ts
@@ -1,7 +1,0 @@
-/**
- * Re-export utils.
- *
- * @Note
- * This negates the need to publish.
- */
-export * from '@aviato/utils'
