--- packages/utils/.gitignore
@@ -1,2 +1,0 @@
-temp
-dist
