--- packages/utils/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@aviato/utils",
-  "version": "1.0.0",
-  "description": "Utils for Aviato",
-  "license": "MIT",
-  "devDependencies": {
-    "eslint-plugin-prettier": "^4.0.0",
-    "rollup": "^2.33.1",
-    "rollup-plugin-delete": "^2.0.0",
-    "rollup-plugin-typescript2": "^0.31.1",
-    "typescript": "4.4.3"
-  },
-  "sideEffects": false,
-  "prettier": "@saulx/prettier-config",
-  "scripts": {
-    "build": "rollup -c",
-    "watch": "rollup -c --watch",
-    "lint": "run-p lint:*",
-    "lint:typescript": "tsc --project ./tsconfig.json --noEmit",
-    "lint:eslint": "eslint --ignore-path .gitignore './**/*.{ts,tsx}'",
-    "lint:prettier": "prettier --write './**/*.ts{,x}'"
-  },
-  "main": "dist/index.js",
-  "module": "dist/index.es.js"
-}
