--- packages/utils/src/index.ts
@@ -1,2 +1,0 @@
-export * from './logging'
-export * from './misc'
