--- packages/utils/src/misc/index.ts
@@ -1,10 +1,0 @@
-/* eslint-disable no-void */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-
-const noop = (...args: any[]): any => void 0
-
-function clamp(value: number, [min, max]: [number, number]): number {
-  return Math.min(max, Math.max(min, value))
-}
-
-export { noop, clamp }
